to seperated method to files and set correct name
