+++ server/controllers/external/index.js
@@ -1,0 +1,13 @@
+import history from './history'
+import realtime from './realtime'
+import objectTable from './objectTable'
+import signIn from './signIn'
+import tracingHistory from './tracingHistory'
+
+export{
+    history,
+    realtime,
+    objectTable,
+    signIn,
+    tracingHistory
+}
